% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph2bitmap.R
\name{graph2bitmap}
\alias{graph2bitmap}
\alias{graph2jpg}
\alias{graph2png}
\alias{graph2tif}
\title{Save currently active R graph to bitmap format}
\usage{
graph2bitmap(file = "Rplot", obj = NULL, fun = NULL, type = c("PNG",
  "JPG", "TIF"), aspectr = NULL, width = NULL, height = NULL, dpi = 600,
  scaling = 100, font = "sans", bg = "white", pngtype = c("cairo-png",
  "windows", "cairo"), tifftype = c("cairo", "windows"),
  tiffcompression = c("lzw", "rle", "jpeg", "zip", "lzw+p", "zip+p"),
  jpegquality = 99, ...)

graph2png(...)

graph2tif(...)

graph2jpg(...)
}
\arguments{
\item{file}{name of output file. Any extension is ignored and added
according to the requested output type. If file already exists it is overwritten.}

\item{obj}{given ggplot2 plot or lattice plot object to export; if
set to NULL the currently active R graph will be exported; not
supported for base R plots.}

\item{fun}{plot passed on as a function used to create it; useful especially
for base R plots.}

\item{type}{desired output type - PNG, TIF or JPG are currently supported.
PNG is the preferred format, as it is a lossless format, and compresses better
than TIF.}

\item{aspectr}{desired width to height aspect ratio. If set to NULL, the
aspect ratio of the graphics device is used. Can also be combined with one
value for either the desired width or height of the graph.}

\item{width}{desired width in inches; can be combined with a desired
aspect ratio aspectr.}

\item{height}{desired height in inches; can be combined with a desired
aspect ratio aspectr.}

\item{dpi}{desired output in dpi; defaults to 600 dpi.}

\item{scaling}{scale width & height by a certain percentage.}

\item{font}{desired font to use for labels; defaults to "sans".}

\item{bg}{desired background colour, e.g. "white" or "transparent".}

\item{pngtype}{use Cairographics for export or Windows GDI?}

\item{tifftype}{use Cairographics for export or Windows GDI?}

\item{tiffcompression}{compression to use for TIF files.}

\item{jpegquality}{quality of JPEG compression.}

\item{\dots}{any other options are passed on to ReporteR's addPlot function.}
}
\description{
Save the currently active R graph or a graph passed as an object or function
to bitmap format with sensible defaults
}
\section{Functions}{
\itemize{
\item \code{graph2png}: 

\item \code{graph2tif}: 

\item \code{graph2jpg}: 
}}
\note{
%% ~~further notes~~
}
\examples{
library(ggplot2)
qplot(Sepal.Length, Petal.Length, data = iris, color = Species, size = Petal.Width, alpha = I(0.7))
graph2png(dpi=1200)
graph2tif()
graph2jpg()
}
\author{
Tom Wenseleers
}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}

