% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph2vector.R
\name{graph2vector}
\alias{graph2eps}
\alias{graph2pdf}
\alias{graph2svg}
\alias{graph2vector}
\title{Save currently active R graph to vector format}
\usage{
graph2vector(file = "Rplot", obj = NULL, fun = NULL, type = "SVG",
  aspectr = NULL, width = NULL, height = NULL, scaling = 100,
  font = "sans", bg = "white", colormodel = "rgb", cairo = TRUE, ...)

graph2svg(...)

graph2pdf(...)

graph2eps(...)
}
\arguments{
\item{file}{name of output file. Any extension is ignored and added
according to the requested output type. If file already exists it is overwritten.}

\item{obj}{given ggplot2 plot or lattice plot object to export; if
set to NULL the currently active R graph will be exported; not
supported for base R plots.}

\item{fun}{plot passed on as a function used to create it; useful especially
for base R plots.}

\item{type}{desired output type - SVG, PDF or EPS are currently supported.
SVG is the preferred format, and good for editing in Inkscape; PDF is good
for printing; EPS is sometimes requested by journals, though lower quality,
especially when semi-transparency is used, as this is rasterized to bitmap.}

\item{aspectr}{desired width to height aspect ratio. If set to NULL, the
aspect ratio of the graphics device is used. Can also be combined with one
value for either the desired width or height of the graph.}

\item{width}{desired width in inches; can be combined with a desired
aspect ratio aspectr.}

\item{height}{desired height in inches; can be combined with a desired
aspect ratio aspectr.}

\item{scaling}{scale width & height by a certain percentage.}

\item{font}{desired font to use for labels; defaults to "sans", which corresponds
to "ArialMT" on Windows systems. Fonts are embedded by default in EPS output.}

\item{bg}{desired background colour, e.g. "white" or "transparent".}

\item{colormodel}{desired colormodel in pdf or eps output when cairo=FALSE;
currently allowed values are "rgb" (default), "cmyk", "srgb", "srgb+gray", "rgb-nogray",
and "gray" (or "grey").}

\item{cairo}{logical indicating whether or not to use the cairo graphics
device for output to PDF or EPS, defaults to TRUE, thereby allowing for
simulated semi-transparency in EPS output, by rasterizing semi-transparent
sections, and automated font embedding.}

\item{\dots}{any other options are passed on to svg, cairo_pdf, cairo_ps, pdf or
postscript.}
}
\description{
Save the currently active R graph or a graph passed as an object or function
to vector format with sensible defaults
}
\section{Functions}{
\itemize{
\item \code{graph2svg}: 

\item \code{graph2pdf}: 

\item \code{graph2eps}: 
}}
\note{
%% ~~further notes~~
}
\examples{
library(ggplot2)
qplot(Sepal.Length, Petal.Length, data = iris, color = Species, size = Petal.Width, alpha = I(0.7))
graph2svg("ggplot2_plot.svg")
graph2pdf("ggplot2_plot.pdf")
graph2eps("ggplot2_plot.eps")
}
\author{
Tom Wenseleers
}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}

