% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plotfit.R
\docType{data}
\name{reqpackagesCRAN}
\alias{plotfit2d}
\alias{plotfit3d}
\alias{reqpackagesCRAN}
\title{Plot predictions of a (general) linear model in 2D or 3D}
\format{An object of class \code{character} of length 8.}
\usage{
reqpackagesCRAN
}
\arguments{
\item{x}{given \code{ggplot2} plot or \code{lattice} plot object to export; if
set to \code{NULL} the currently active R graph will be exported; not
supported for base R plots.}

\item{file}{name of output file. Any extension is ignored and added
according to the requested output type. If file already exists it is overwritten.}

\item{fun}{plot passed on as a function used to create it; useful especially
for base R plots.}

\item{type}{desired output type - \code{PNG}, \code{TIF} or \code{JPG} are currently supported.
\code{PNG} is the preferred format, as it is a lossless format, and compresses better
than \code{TIF}.}

\item{aspectr}{desired width to height aspect ratio. If set to \code{NULL}, the
aspect ratio of the graphics device is used. Can also be combined with one
value for either the desired width or height of the graph.}

\item{width}{desired width in inches; can be combined with a desired
aspect ratio aspectr.}

\item{height}{desired height in inches; can be combined with a desired
aspect ratio aspectr.}

\item{dpi}{desired output in dpi; defaults to 600 dpi.}

\item{scaling}{scale width & height by a certain percentage.}

\item{font}{desired font to use for labels in PNG and TIFF output; defaults to 
\code{"Arial"} on Windows systems and to \code{"Helvetica"} on other systems.}

\item{bg}{desired background colour, e.g. \code{"white"} or \code{"transparent"}.}

\item{cairo}{logical, specifying whether or not to use \code{Cairographics} for export.}

\item{tiffcompression}{compression to use for \code{TIF} files.}

\item{jpegquality}{quality of \code{JPEG} compression.}

\item{\dots}{any other options are passed on to \code{ReporteRs}'s \code{\link[ReporteRs]{addPlot}} function.}
}
\value{
\code{NULL}
}
\description{
Plot the predictions of a [general(ized)] linear model as a colour-coded 
2D contour plot or 3D surface, if desired together with the raw data
}
\examples{
library(ggplot2)
x=qplot(Sepal.Length, Petal.Length, data = iris, 
        color = Species, size = Petal.Width, alpha = I(0.7))
x
graph2png(x=x,dpi=400)
graph2tif(x=x)
graph2jpg(x=x)

\dontrun{
# use active graph instead of passing plot as object
graph2png(aspectr=1.7)
}
}
\author{
Tom Wenseleers
}
\seealso{
\code{\link{graph2office}}, \code{\link{graph2vector}}, \code{\link{graph2svg}}, \code{\link{graph2pdf}},
\code{\link{graph2eps}}, \code{\link{graph2tex}}
}
\keyword{datasets}

