% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/table2ppt.R
\name{table2ppt}
\alias{table2ppt}
\title{Export statistical output to a table in Microsoft Powerpoint / LibreOffice format}
\usage{
table2ppt(obj = NULL, file = "Rtable", append = FALSE, digits = 2,
  digitspvals = 2, width = NULL, height = NULL, offx = 1, offy = 1,
  font = "Calibri", pointsize = 12, add.rownames = FALSE, ...)
}
\arguments{
\item{obj}{given R stats object to export; if set to NULL the output of the
previous R command will be exported.}

\item{file}{name of output file. The .pptx extension is added automatically.}

\item{append}{logical value - if TRUE and type="PPT"" or "DOC"" it will
append the graph to the given file, where file can also be a given corporate
template in case type="PPT".  If append=FALSE any existing file will be
overwritten.}

\item{digits}{number of significant digits to show for all columns except
for the column with p values.}

\item{digitspvals}{number of significant digits to show for columns with p
values.}

\item{width}{desired width of table in inches.}

\item{height}{desired height of table in inches.}

\item{offx}{x offset in inches to specify horizontal location of table.}

\item{offy}{y offset in inches to specify vertical location of table.}

\item{font}{desired font to use for output table.}

\item{pointsize}{desired font point size.}

\item{add.rownames}{logical specifying whether or not to add row names.}

\item{\dots}{extra options are passed on to xtable.}
}
\value{
FlexTable object
}
\description{
Export currently showing R stats object or stats object obj to a Microsoft
Powerpoint / LibreOffice table
}
\note{
Columns corresponding to degrees of freedom (with header "Df" or "df")
are always given as integers.
}
\examples{
# export of aov Anova output
fit=aov(yield ~ block + N * P + K, npk)
summary(fit)
table2ppt(file="table_aov.pptx")  
summary(fit)
table2ppt(file="table_aov.pptx", width=9, append=TRUE) # append table to previous slide
summary(fit)
table2ppt(file="table_aov.docx", font="Times New Roman", pointsize=14, 
          digits=4, digitspvals=1, append=TRUE)
}
\author{
Tom Wenseleers
}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}

