% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/table2tex.R
\name{table2tex}
\alias{table2doc}
\alias{table2html}
\alias{table2tex}
\alias{table2tex2}
\title{Export statistical output to a Microsoft Word / LibreOffice, HTML or Latex table}
\usage{
table2tex(file = "Rtable", obj = NULL, type = "TEX", digits = 2,
  summary = FALSE, standAlone = TRUE, ...)

table2tex2(...)

table2doc(...)

table2html(...)
}
\arguments{
\item{file}{name of output file. The appropriate extension is added automatically.}

\item{obj}{given R stats object or list of stats objects to export; if set to NULL
the output of the previous R command will be exported.}

\item{digits}{number of significant digits to show. A value of NA indicates that
no rounding should be done.}

\item{summary}{logical indicating whether or not to summarize data files.}

\item{\dots}{extra options are passed on to stargazer.}
}
\description{
Export currently showing R stats object or stats object obj to a
Microsoft Word / LibreOffice, HTML or Latex table
}
\section{Functions}{
\itemize{
\item \code{table2tex2}: 

\item \code{table2doc}: 

\item \code{table2html}: 
}}
\note{
Objects that can be exported are all those supported by stargazer,
namely aftreg (eha), anova (stats), aov (stats), aovlist (stats), arima (stats),
}
\examples{
# export of aov Anova output
fit=aov(yield ~ block + N * P + K, npk)
summary(fit)
table2tex(file="table_aov.tex") # export to Latex in standAlone format
summary(fit)
table2tex2(file="table_aov.tex") # export to Latex to paste in document
summary(fit)
table2doc(file="table_aov.doc") # export to Word
table2doc(file="table_aov.doc",obj=fit) # export to Word, but pass as object
summary(fit)
table2html(file="table_aov.html") # export to HTML
}
\author{
Tom Wenseleers
}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}

