% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph2bitmap.R
\name{graph2bitmap}
\alias{graph2bitmap}
\alias{graph2jpg}
\alias{graph2png}
\alias{graph2tif}
\title{Save currently active R graph to bitmap format}
\usage{
graph2bitmap(obj = NULL, file = "Rplot", fun = NULL, type = c("PNG",
  "JPG", "TIF"), aspectr = NULL, width = NULL, height = NULL, dpi = 600,
  scaling = 100, font = ifelse(Sys.info()["sysname"] == "Windows", "Arial",
  "Helvetica")[[1]], bg = "white", pngtype = c("cairo-png", "windows",
  "cairo"), tifftype = c("cairo", "windows"), tiffcompression = c("lzw",
  "rle", "jpeg", "zip", "lzw+p", "zip+p"), jpegquality = 99, ...)

graph2png(...)

graph2tif(...)

graph2jpg(...)
}
\arguments{
\item{obj}{given \code{ggplot2} plot or \code{lattice} plot object to export; if
set to \code{NULL} the currently active R graph will be exported; not
supported for base R plots.}

\item{file}{name of output file. Any extension is ignored and added
according to the requested output type. If file already exists it is overwritten.}

\item{fun}{plot passed on as a function used to create it; useful especially
for base R plots.}

\item{type}{desired output type - \code{PNG}, \code{TIF} or \code{JPG} are currently supported.
\code{PNG} is the preferred format, as it is a lossless format, and compresses better
than \code{TIF}.}

\item{aspectr}{desired width to height aspect ratio. If set to \code{NULL}, the
aspect ratio of the graphics device is used. Can also be combined with one
value for either the desired width or height of the graph.}

\item{width}{desired width in inches; can be combined with a desired
aspect ratio aspectr.}

\item{height}{desired height in inches; can be combined with a desired
aspect ratio aspectr.}

\item{dpi}{desired output in dpi; defaults to 600 dpi.}

\item{scaling}{scale width & height by a certain percentage.}

\item{font}{desired font to use for labels; defaults to \code{"Arial"} on Windows
systems and to \code{"Helvetica"} on other systems.}

\item{bg}{desired background colour, e.g. \code{"white"} or \code{"transparent"}.}

\item{pngtype}{use \code{Cairographics} for export or Windows GDI?}

\item{tifftype}{use \code{Cairographics} for export or Windows GDI?}

\item{tiffcompression}{compression to use for \code{TIF} files.}

\item{jpegquality}{quality of \code{JPEG} compression.}

\item{\dots}{any other options are passed on to \code{ReporteRs}'s \code{\link[ReporteRs]{addPlot}} function.}
}
\value{
\code{NULL}
}
\description{
Save the currently active R graph or a graph passed as an object or function
to bitmap format with sensible defaults
}
\section{Functions}{
\itemize{
\item \code{graph2png}: 

\item \code{graph2tif}: 

\item \code{graph2jpg}: 
}}
\examples{
library(ggplot2)
qplot(Sepal.Length, Petal.Length, data = iris, color = Species, size = Petal.Width, alpha = I(0.7))
graph2png(dpi=1200)
graph2tif()
graph2jpg()
}
\author{
Tom Wenseleers
}
\seealso{
\code{\link{graph2office}}, \code{\link{graph2vector}}, \code{\link{graph2svg}}, \code{\link{graph2pdf}},
\code{\link{graph2eps}}, \code{\link{graph2tex}}
}

