% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph2tex.R
\name{graph2tex}
\alias{graph2tex}
\alias{graph2tex2}
\title{Save currently active R graph to Latex format}
\usage{
graph2tex(file = "Rplot", obj = NULL, fun = NULL, aspectr = NULL,
  width = NULL, height = NULL, scaling = 100, font = "sans",
  bg = "transparent", standAlone = TRUE, ...)

graph2tex2(...)
}
\arguments{
\item{file}{name of output file; extension ".tex" is added automatically.
If file already exists it is overwritten.}

\item{obj}{given ggplot2 plot or lattice plot object to export; if
set to NULL the currently active R graph will be exported; not
supported for base R plots.}

\item{fun}{plot passed on as a function used to create it; useful especially
for base R plots.}

\item{aspectr}{desired width to height aspect ratio. If set to NULL, the
aspect ratio of the graphics device is used. Can also be combined with one
value for either the desired width or height of the graph.}

\item{width}{desired width in inches; can be combined with a desired
aspect ratio aspectr.}

\item{height}{desired height in inches; can be combined with a desired
aspect ratio aspectr.}

\item{scaling}{scale width & height by a certain percentage.}

\item{font}{default "sans" uses sansmath font for output; any other value
will result in the font of the Latex document being used.}

\item{bg}{desired background colour, e.g. "white" or "transparent".}

\item{standAlone}{logical indicating whether output should be a full compilable
Latex document or whether only the graph output should be produced, for pasting
into another document. graph2tex default is standAlone=TRUE, whereas graph2tex2
has default standAlone=FALSE.}

\item{\dots}{any other options are passed on to tikzDevice's tikz function.}
}
\description{
Save the currently active R graph or a graph passed as an object or function
to Latex format with sensible defaults
}
\section{Functions}{
\itemize{
\item \code{graph2tex2}: 
}}
\note{
%% ~~further notes~~
}
\examples{
\dontrun{
library(ggplot2)
qplot(Sepal.Length, Petal.Length, data = iris, color = Species, 
      size = Petal.Width, alpha = I(0.7))
# this results in a compilable Latex file with sans font
graph2tex("ggplot2_plot1.tex")
# this results in a compilable Latex file with default document font
graph2tex("ggplot2_plot2.tex",font=NA)
# this results in Latex output that has to be pasted into another document
graph2tex2("ggplot2_plot3.tex",font=NA)
}
}
\author{
Tom Wenseleers
}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}

