% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2tex.R
\name{table2tex}
\alias{table2doc}
\alias{table2html}
\alias{table2tex}
\alias{table2tex2}
\title{Export statistical output to a Microsoft Word / LibreOffice, HTML or Latex table}
\usage{
table2tex(x = NULL, file = "Rtable", type = "TEX", digits = 2,
  summary = FALSE, standAlone = TRUE, ...)
}
\arguments{
\item{x}{given R stats object or list of stats objects to export; if set to \code{NULL} 
the output of the previous R command will be exported.}

\item{file}{name of output file. The appropriate extension is added automatically.}

\item{type}{desired output type - \code{"TEX"} for Latex, \code{"DOC"} for Word and \code{"HTML"} for HTML.}

\item{digits}{number of significant digits to show. A value of NA indicates that
no rounding should be done.}

\item{summary}{logical indicating whether or not to summarize data files.}

\item{standAlone}{logical indicating whether exported Latex code should be
standalone compilable, or whether it will be pasted into another document.}

\item{\dots}{extra options are passed on to stargazer.}
}
\value{
\code{NULL}
}
\description{
Export currently showing R stats object or stats object obj to a
Microsoft Word / LibreOffice, HTML or Latex table
}
\details{
Objects that can be exported are all those supported by \code{\link[stargazer]{stargazer}},
namely \code{aftreg} (\code{eha}), \code{anova} (\code{stats}), \code{aov} (\code{stats}), \code{aovlist} (\code{stats}), \code{arima} (\code{stats}), 
\code{betareg} (\code{betareg}), \code{binaryChoice} (\code{sampleSelection}), \code{bj} (\code{rms}), \code{brglm} (\code{brglm}), 
\code{censReg} (\code{censReg}), \code{coeftest} (\code{lmtest}), \code{coxph} (\code{survival}), \code{coxreg} (\code{eha}), 
\code{clm} (\code{ordinal}), \code{clogit} (\code{survival}), \code{cph} (\code{rms}), \code{dynlm} (\code{dynlm}), \code{ergm} (\code{ergm}), 
\code{errorsarlm} (\code{spdev}), \code{felm} (\code{lfe}), \code{gam} (\code{mgcv}), \code{garchFit} (\code{fGarch}), \code{gee} (\code{gee}), 
\code{glm} (\code{stats}), \code{Glm} (\code{rms}), \code{glmer} (\code{lme4}), \code{glmrob} (\code{robustbase}), \code{gls} (\code{nlme}), 
\code{Gls} (\code{rms}), \code{gmm} (\code{gmm}), \code{heckit} (\code{sampleSelection}), \code{hetglm} (\code{glmx}), \code{hurdle} (\code{pscl}), 
\code{ivreg} (\code{AER}), \code{lagarlm} (\code{spdep}), \code{lm} (\code{stats}), \code{lme} (\code{nlme}), \code{lmer} (\code{lme4}), \code{lmrob} (\code{robustbase}), 
\code{lrm} (\code{rms}), \code{maBina} (\code{erer}), \code{mclogit} (\code{mclogit}), \code{mlogit} (\code{mlogit}), \code{mnlogit} (\code{mnlogit}), 
\code{mlreg} (\code{eha}), \code{multinom} (\code{nnet}), \code{nlme} (\code{nlme}), \code{nlmer} (\code{lme4}), \code{ols} (\code{rms}), \code{pgmm} (\code{plm}), 
\code{phreg} (\code{eha}), \code{plm} (\code{plm}), \code{pmg} (\code{plm}), \code{polr} (\code{MASS}), \code{psm} (\code{rms}), \code{rem.dyad} (\code{relevent}), 
\code{rlm} (\code{MASS}), \code{rq} (\code{quantreg}), \code{Rq} (\code{rms}), \code{selection} (\code{sampleSelection}), \code{svyglm} (\code{survey}), 
\code{survreg} (\code{survival}), \code{tobit} (\code{AER}), \code{weibreg} (\code{eha}), \code{zeroin} (\code{pscl}), \code{relogit} (\code{zelig}), 
\code{cloglog.net} (\code{zelig}), \code{gamma.net} (\code{zelig}), \code{probit.net} (\code{zelig}) and \code{logit.net} (\code{zelig}).
}
\examples{
# export of aov Anova output
fit=aov(yield ~ block + N * P + K, npk)
x=summary(fit)
table2tex(x=x,file="table_aov1.tex") # export to Latex in standAlone format
table2tex2(x=x,file="table_aov2.tex") # export to Latex to paste in document
table2doc(x=x,file="table_aov.doc") # export to Word
table2html(x=x,file="table_aov.html") # export to HTML

\dontrun{
# export output of last evaluation
summary(fit)
table2doc(file="table_aov.doc") # export to Word
}
}
\author{
Tom Wenseleers
}
\seealso{
\code{\link{table2ppt}}
}

